+++ packages/reference/src/resources/SingleResourceReferenceEditor.tsx
@@ -16,11 +16,9 @@ export function SingleResourceReferenceEditor(
     apiUrl: string;
   }
 ) {
-  const { dialogs, field } = props.sdk;
   const linkActionsProps = useResourceLinkActions({
-    dialogs,
-    field,
-    apiUrl: props.apiUrl,
+    sdk: props.sdk,
+    parameters: props.parameters,
   });
 
   return (