+++ packages/reference/src/resources/useResourceLinkActions.ts
@@ -1,8 +1,9 @@
 import { useMemo } from 'react';
 
-import type { FieldAPI, FieldAppSDK } from '@contentful/app-sdk';
+import type { FieldAPI } from '@contentful/app-sdk';
 import type { ResourceLink } from 'contentful-management';
 
+import { EditorPermissionsProps, useEditorPermissions } from '../common/useEditorPermissions';
 import { LinkActionsProps } from '../components';
 
 const getUpdatedValue = (
@@ -18,12 +19,14 @@ const getUpdatedValue = (
   }
 };
 
+type ResourceLinkActionProps = Pick<EditorPermissionsProps, 'parameters' | 'sdk'>;
+
 export function useResourceLinkActions({
-  dialogs,
-  field,
-}: Pick<FieldAppSDK, 'field' | 'dialogs'> & {
-  apiUrl: string;
-}): LinkActionsProps {
+  parameters,
+  sdk,
+}: ResourceLinkActionProps): LinkActionsProps {
+  const { field, dialogs } = sdk;
+
   const onLinkedExisting = useMemo(() => {
     return (
       links: ResourceLink<'Contentful:Entry'>[] | [ResourceLink<'Contentful:Entry'> | null]
@@ -58,6 +61,13 @@ export function useResourceLinkActions({
     // @ts-expect-error wait for update of app-sdk version
   }, [dialogs, field.allowedResources, multiple, onLinkedExisting]);
 
+  const { canLinkEntity } = useEditorPermissions({
+    entityType: 'Entry',
+    allContentTypes: [],
+    sdk,
+    parameters,
+  });
+
   return {
     onLinkExisting,
     // @ts-expect-error
@@ -67,7 +77,7 @@ export function useResourceLinkActions({
     contentTypes: [],
     canCreateEntity: false,
     canLinkMultiple: multiple,
-    canLinkEntity: true,
+    canLinkEntity,
     isDisabled: false,
     isEmpty: false,
     isFull: false,