+++ packages/rich-text/src/plugins/shared/FetchingWrappedResourceCard.tsx
@@ -44,7 +44,6 @@ const InternalEntryCard = React.memo((props: InternalEntryCard) => {
       isClickable={false}
       getEntityScheduledActions={() => Promise.resolve([])}
       useLocalizedEntityStatus={props.sdk.parameters.instance.useLocalizedEntityStatus}
-      isLocalized={!!('localized' in props.sdk.field && props.sdk.field.localized)}
     />
   );
 }, areEqual);