+++ packages/json/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.3.35](https://github.com/contentful/field-editors/compare/@contentful/field-editor-json@3.3.34...@contentful/field-editor-json@3.3.35) (2025-01-17)
+
+**Note:** Version bump only for package @contentful/field-editor-json
+
 ## [3.3.34](https://github.com/contentful/field-editors/compare/@contentful/field-editor-json@3.3.33...@contentful/field-editor-json@3.3.34) (2024-11-26)
 
 **Note:** Version bump only for package @contentful/field-editor-json